Reduce size
git-svn-id: https://svn.code.sf.net/p/tass64/code/trunk@2550 111d7451-93fa-4e81-90d4-8039f76616e9

+++ unicode.h
@@ -25,7 +25,7 @@
 
 struct ubuff_s {
     uchar_t *data;
-    size_t len, p;
+    uint32_t len, p;
 };
 
 struct str_t;
